test: after_initial_init_emits_idle_ready_signal
This required me to pull in a published event recorder, which is a test support utility provided by my cpputest-for-qpcpp library.

We still aren't doing anything real, but that
is about to change, as the logical next
test is to start up the CLI and see if it
writes anything to the output.