the_service_supports_configurable_max_cli_binding_count (false passing)
In this case the test is passing, but it really should be failing, as the test should be configuring the service for a max CLI count of 1, yet submitting three seems to work without any errors or asserts.
I'm committing this to demonstrate a key part of the TDD process:  mentally always have a hypothesis before building/running tests. Do the results match? If not, why not?